+++ include/config.h
@@ -3,8 +3,8 @@
  
 #define MODULE_NAME "bscan"
 #define VERSION_MAJOR "1"
-#define VERSION_MINOR "6"
-#define VERSION_PATCH "0.1.6"
-#define MODULE_VERSION "0.1.6"
+#define VERSION_MINOR "7"
+#define VERSION_PATCH "0.1.7"
+#define MODULE_VERSION "0.1.7"
  
 #endif // CONFIG_H_IN